feat(layout): add flex rules push footer down
Gave the container the display of flex and set its directoin to column
in order to stack the .header, .content, and .footer vertically. Giving
the min-height of 100vh was also done to make sure the container always
takes at least the full viewport height.

Additionally, flex: 1 was added as a rule to .content to push the footer
down when there's extra space provided. This pushed it down to the very
bottom of the web page.